+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Features
 
+- Add new --no-extension flag, see #673 (@jacobmischka)
 - Improved the usability of the time-based options, see #624 and #645 (@gorogoroumaru)
 - Add new `--prune` flag, see #535 (@reima)